requirements.txt
absl-py==1.0.0
black==21.12b0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
cycler==0.11.0
fonttools==4.28.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.42.0
idna==3.3
importlib-metadata==4.8.2
kiwisolver==1.3.2
Markdown==3.3.6
matplotlib==3.5.0
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mypy-extensions==0.4.3
numpy==1.21.2
oauthlib==3.1.1
olefile==0.46
packaging==21.3
pathspec==0.9.0
Pillow==8.4.0
pip==21.2.4
platformdirs==2.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
python-dateutil==2.8.2
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
setuptools==58.0.4
setuptools-scm==6.3.2
six==1.16.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tomli==1.2.2
torch==1.10.0
torchaudio==0.10.0
torchvision==0.11.1
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
wheel==0.37.0
zipp==3.6.0
